Converting async query tests to use AssertQuery verification infrastructure
- creating AsyncQueryTestBase with all the async helper methods,
- adapting ResultAsserter to support async calls,
- removing static NorthwindData, since it's no longer used
