+++ src/EFCore.Specification.Tests/Query/AsyncQueryNavigationsTestBase.cs
@@ -1,25 +1,26 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore.TestModels.Northwind;
 using Microsoft.EntityFrameworkCore.TestUtilities;
 using Microsoft.EntityFrameworkCore.TestUtilities.Xunit;
-using Xunit;
 
 // ReSharper disable InconsistentNaming
 namespace Microsoft.EntityFrameworkCore.Query
 {
-    public abstract class AsyncQueryNavigationsTestBase<TFixture> : IClassFixture<TFixture>
+    public abstract class AsyncQueryNavigationsTestBase<TFixture> : AsyncQueryTestBase<TFixture>
         where TFixture : NorthwindQueryFixtureBase<NoopModelCustomizer>, new()
     {
-        protected AsyncQueryNavigationsTestBase(TFixture fixture) => Fixture = fixture;
 
-        protected TFixture Fixture { get; }
-        
+        protected AsyncQueryNavigationsTestBase(TFixture fixture)
+            : base(fixture)
+        {
+        }
+
+        protected NorthwindContext CreateContext() => Fixture.CreateContext();
+
         [ConditionalFact]
         public virtual async Task Include_with_multiple_optional_navigations()
         {
@@ -39,36 +40,6 @@ namespace Microsoft.EntityFrameworkCore.Query
                     .Include(od => od.Product)
                     .Where(od => od.Order.Customer.City == "London"),
                 entryCount: 221);
-        }
-
-        protected NorthwindContext CreateContext() => Fixture.CreateContext();
-
-        protected async Task AssertQuery<TItem>(
-            Func<IQueryable<TItem>, IQueryable<object>> query,
-            bool assertOrder = false,
-            int entryCount = 0,
-            Action<IList<object>, IList<object>> asserter = null)
-            where TItem : class
-            => await AssertQuery(query, query, assertOrder, entryCount, asserter);
-
-        protected async Task AssertQuery<TItem>(
-            Func<IQueryable<TItem>, IQueryable<object>> efQuery,
-            Func<IQueryable<TItem>, IQueryable<object>> l2oQuery,
-            bool assertOrder = false,
-            int entryCount = 0,
-            Action<IList<object>, IList<object>> asserter = null)
-            where TItem : class
-        {
-            using (var context = CreateContext())
-            {
-                TestHelpers.AssertResults(
-                    l2oQuery(NorthwindData.Set<TItem>()).ToArray(),
-                    await efQuery(context.Set<TItem>()).ToArrayAsync(),
-                    assertOrder,
-                    asserter);
-
-                Assert.Equal(entryCount, context.ChangeTracker.Entries().Count());
-            }
         }
     }
 }
